+++ linbox/tests/test-diagonal.C
@@ -281,7 +281,7 @@ int main (int argc, char **argv)
 		{ 'i', "-i I", "Perform each test for I iterations (default 100)",          TYPE_INT,     &iterations },
 	};
 
-	typedef Modular<long> Field;
+	typedef Modular<uint32> Field;
 
 	parseArguments (argc, argv, args);
 	Field F (q);
